rename fileFormatter to tabularFileFormatter
